+++ tests/managers/test_managers_inherited.py
@@ -54,7 +54,7 @@ class User(BaseModel):
 
 class Product(BaseModel):
     name = edgy.CharField(max_length=100)
-    rating = edgy.IntegerField(minimum=1, maximum=5)
+    rating = edgy.IntegerField(gte=1, lte=5)
     in_stock = edgy.BooleanField(default=False)
     is_active = edgy.BooleanField(default=False)