+++ tests/uniques/test_unique_constraint.py
@@ -40,7 +40,7 @@ class User(BaseModel):
 class HubUser(BaseModel):
     name = edgy.CharField(max_length=255)
     email = edgy.CharField(max_length=60, null=True)
-    age = edgy.IntegerField(minimum=18, null=True)
+    age = edgy.IntegerField(gte=18, null=True)
 
     class Meta:
         unique_together = [