+++ tests/uniques/test_unique_constraint_inherited.py
@@ -46,7 +46,7 @@ class User(AbsUser, BaseModel):
 class AbsHubUser(edgy.StrictModel):
     name = edgy.CharField(max_length=255)
     email = edgy.CharField(max_length=60, null=True)
-    age = edgy.IntegerField(minimum=18, null=True)
+    age = edgy.IntegerField(gte=18, null=True)
 
     class Meta:
         abstract = True